umsh_cli::events

Enum CliEvent

pub enum CliEvent {

    Received {
        from: PublicKey,
        hops: u8,
        rssi: i16,
        snr: i16,
        prefix: Vec<u8, EVENT_PAYLOAD_MAX>,
    },
    AckReceived {
        peer: PublicKey,
    },
    AckTimeout {
        peer: PublicKey,
    },
    NodeDiscovered {
        from: PublicKey,
        name: Option<String<32>>,
    },
    Beacon {
        hint: NodeHint,
        from: Option<PublicKey>,
    },
    PfsEstablished {
        peer: PublicKey,
    },
    PfsEnded {
        peer: PublicKey,
    },
    PfsFailed {
        peer: PublicKey,
        reason: PfsFailure,
    },
    Pong {
        peer: PublicKey,
        rtt_ms: u64,
    },
    PingTimeout {
        peer: PublicKey,
    },
    UnknownMacCmdIn {
        peer: PublicKey,
        cmd_id: u8,
    },
    RawTx {
        bytes: Vec<u8, EVENT_RAW_MAX>,
    },
    RawRx {
        bytes: Vec<u8, EVENT_RAW_MAX>,
    },
    OutputLine {
        line: String<EVENT_LINE_MAX>,
    },
    SendText {
        peer: PublicKey,
        body: String<EVENT_LINE_MAX>,
    },
    StartPfs {
        peer: PublicKey,
        minutes: u16,
    },
    EndPfs {
        peer: PublicKey,
    },
    ChannelSend {
        channel_slot: u8,
        body: String<EVENT_LINE_MAX>,
    },
    SendBeacon,
    SendRaw {
        peer: PublicKey,
        bytes: Vec<u8, EVENT_PAYLOAD_MAX>,
    },

}

Variants

Received

Fields

from: PublicKey

hops: u8

rssi: i16

snr: i16

prefix: Vec<u8, EVENT_PAYLOAD_MAX>

AckReceived

Fields

peer: PublicKey

AckTimeout

Fields

peer: PublicKey

NodeDiscovered

Fields

from: PublicKey

name: Option<String<32>>

Beacon

Fields

hint: NodeHint

3-byte source hint from the packet header — always present.

from: Option<PublicKey>

Full source pubkey, when the beacon carried it (some senders only include the hint to save airtime).

PfsEstablished

Fields

peer: PublicKey

PfsEnded

Fields

peer: PublicKey

PfsFailed

Fields

peer: PublicKey

reason: PfsFailure

Pong

Fields

peer: PublicKey

rtt_ms: u64

PingTimeout

Fields

peer: PublicKey

UnknownMacCmdIn

Fields

peer: PublicKey

cmd_id: u8

RawTx

Fields

bytes: Vec<u8, EVENT_RAW_MAX>

RawRx

Fields

bytes: Vec<u8, EVENT_RAW_MAX>

OutputLine

Fields

line: String<EVENT_LINE_MAX>

SendText

Fields

peer: PublicKey

body: String<EVENT_LINE_MAX>

StartPfs

Fields

peer: PublicKey

minutes: u16

EndPfs

Fields

peer: PublicKey

ChannelSend

Fields

channel_slot: u8

body: String<EVENT_LINE_MAX>

SendBeacon

SendRaw

Fields

peer: PublicKey

bytes: Vec<u8, EVENT_PAYLOAD_MAX>